Separate display function to different js files
